requirements.txt
transformers>=4.40.0
accelerate>=0.28.0
safetensors>=0.4.2
torch>=2.2.0
ollama>=0.1.9
modelcontextprotocol>=0.1.0
pydantic>=2.6.0
typer>=0.12.0
pyyaml>=6.0.1
rich>=13.7.0